Extract into context
