+++ packages/vfx-composer-r3f/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./context"
 export * from "./Emitter"
 export * from "./hooks/particles"
 export * from "./InstancedParticles"
